refactor (separate package) : bounded context에 맞게 패키지 쪼개기 menu
