+++ src/main/java/kitchenpos/menu/model/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.model;
+package kitchenpos.menu.model;
 
 public class MenuGroup {
     private Long id;
