+++ src/test/java/kitchenpos/menu/MenuGroupBoTest.java
@@ -1,7 +1,8 @@
-package kitchenpos.bo;
+package kitchenpos.menu;
 
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.model.MenuGroup;
+import kitchenpos.menu.dao.MenuGroupDao;
+import kitchenpos.menu.bo.MenuGroupBo;
+import kitchenpos.menu.model.MenuGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
