+++ src/test/java/kitchenpos/menu/controller/MenuRestControllerTest.java
@@ -1,7 +1,7 @@
-package kitchenpos.controller;
+package kitchenpos.menu.controller;
 
-import kitchenpos.bo.MenuGroupBo;
-import kitchenpos.model.MenuGroup;
+import kitchenpos.menu.bo.MenuBo;
+import kitchenpos.menu.model.Menu;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.web.servlet.HttpEncodingAutoConfiguration;
@@ -15,7 +15,7 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import java.util.Arrays;
 
-import static kitchenpos.Fixtures.twoChickens;
+import static kitchenpos.Fixtures.twoFriedChickens;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -23,24 +23,25 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@WebMvcTest(MenuGroupRestController.class)
+@WebMvcTest(MenuRestController.class)
 @Import(HttpEncodingAutoConfiguration.class)
-class MenuGroupRestControllerTest {
+class MenuRestControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
     @MockBean
-    private MenuGroupBo menuGroupBo;
+    private MenuBo menuBo;
 
     @Test
     void create() throws Exception {
         // given
-        given(menuGroupBo.create(any(MenuGroup.class))).willReturn(twoChickens());
+        given(menuBo.create(any(Menu.class))).willReturn(twoFriedChickens());
 
         // when
-        final ResultActions resultActions = mockMvc.perform(post("/api/menu-groups")
+        final ResultActions resultActions = mockMvc.perform(post("/api/menus")
                 .contentType(MediaType.APPLICATION_JSON)
-                .content("{\"name\":\"두마리메뉴\"}")
+                .content("{\"name\":\"후라이드+후라이드\",\"price\":19000,\"menuGroupId\":1,"
+                        + "\"menuProducts\":[{\"productId\":1,\"quantity\":2}]}")
         );
 
         // then
@@ -49,16 +50,21 @@ class MenuGroupRestControllerTest {
                 .andExpect(header().exists(HttpHeaders.LOCATION))
                 .andExpect(jsonPath("$.id").isNumber())
                 .andExpect(jsonPath("$.name").isString())
+                .andExpect(jsonPath("$.price").isNumber())
+                .andExpect(jsonPath("$.menuGroupId").isNumber())
+                .andExpect(jsonPath("$.menuProducts").isArray())
+                .andExpect(jsonPath("$.menuProducts[0].productId").isNumber())
+                .andExpect(jsonPath("$.menuProducts[0].quantity").isNumber())
         ;
     }
 
     @Test
     void list() throws Exception {
         // given
-        given(menuGroupBo.list()).willReturn(Arrays.asList(twoChickens()));
+        given(menuBo.list()).willReturn(Arrays.asList(twoFriedChickens()));
 
         // when
-        final ResultActions resultActions = mockMvc.perform(get("/api/menu-groups"));
+        final ResultActions resultActions = mockMvc.perform(get("/api/menus"));
 
         // then
         resultActions.andDo(print())
@@ -66,6 +72,11 @@ class MenuGroupRestControllerTest {
                 .andExpect(jsonPath("$").isArray())
                 .andExpect(jsonPath("$[0].id").isNumber())
                 .andExpect(jsonPath("$[0].name").isString())
+                .andExpect(jsonPath("$[0].price").isNumber())
+                .andExpect(jsonPath("$[0].menuGroupId").isNumber())
+                .andExpect(jsonPath("$[0].menuProducts").isArray())
+                .andExpect(jsonPath("$[0].menuProducts[0].productId").isNumber())
+                .andExpect(jsonPath("$[0].menuProducts[0].quantity").isNumber())
         ;
     }
 }
